pyproject.toml
[project]
name = "deep-ecg-2025"
version = "0.1.0"
description = "Bachelor work, NSU, 2025"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dvc-s3>=3.2.2",
    "hydra-core>=1.3.2",
    "lightning>=2.6.0",
    "mlflow>=3.8.1",
    "onnx>=1.20.1",
    "onnxruntime>=1.23.2",
    "torch>=2.9.1",
    "boto3>=1.42.19",
    "s3fs>=2026.1.0",
    "wfdb>=4.3.0",
    "onnxscript>=0.5.7",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "pre-commit>=4.5.1",
]
